WIP clean up main ( split to mainBase)
