+++ app/Main.hs
@@ -13,7 +13,6 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 
 module Main 
-
   where 
 
 import Prelude ()
@@ -101,172 +100,8 @@ import Debug.Trace
 debug = flip Debug.Trace.trace
 
 
-data Version = Version 
-  { _version :: String 
-  } deriving (Eq, Show, Generic)
-
-$(deriveJSON defaultOptions ''Version)
-instance ToSchema Version
-
 version1 :: Version 
 version1 = Version "0.42.12"
-
-
-
-data DealType = MDeal (DB.TestDeal AB.Mortgage)
-              | LDeal (DB.TestDeal AB.Loan)
-              | IDeal (DB.TestDeal AB.Installment) 
-              | RDeal (DB.TestDeal AB.Lease) 
-              | FDeal (DB.TestDeal AB.FixedAsset) 
-              | VDeal (DB.TestDeal AB.Receivable)
-              | PDeal (DB.TestDeal AB.ProjectedCashflow) 
-              | UDeal (DB.TestDeal AB.AssetUnion)
-              deriving(Show, Generic)
-
-makePrisms ''DealType
-
-instance ToSchema CF.CashFlowFrame
-instance ToSchema AB.Loan
-instance ToSchema AB.Installment
-instance ToSchema AB.AccrualPeriod
-instance ToSchema AB.LeaseStepUp
-instance ToSchema AB.Lease
-instance ToSchema AB.FixedAsset
-instance ToSchema AB.Receivable
-instance ToSchema AB.ProjectedCashflow
-instance ToSchema CutoffFields
-instance ToSchema (P.Pool AB.Mortgage)
-instance ToSchema (P.Pool AB.Loan)
-instance ToSchema (P.Pool AB.Installment)
-instance ToSchema (P.Pool AB.Lease)
-instance ToSchema (P.Pool AB.FixedAsset)
-instance ToSchema (P.Pool AB.Receivable)
-instance ToSchema (P.Pool AB.AssetUnion)
-instance ToSchema (P.Pool AB.ProjectedCashflow)
-instance ToSchema AB.AssetUnion
-instance ToSchema PoolId
-instance ToSchema DealStatus
-instance ToSchema DateType
-instance ToSchema DB.DateDesp
-instance ToSchema DB.ActionOnDate
-instance ToSchema DealStats
-instance ToSchema Cmp
-instance ToSchema PricingMethod
-instance ToSchema Stmt.TxnComment
-instance ToSchema BookDirection
-instance ToSchema Limit
-instance ToSchema PoolSource
-instance ToSchema (RoundingBy Rate)
-instance ToSchema (RoundingBy Integer)
-instance ToSchema (RoundingBy Balance)
-instance ToSchema DealCycle
-instance ToSchema (Table Balance Balance)
-instance ToSchema (Table Float Spread)
-instance ToSchema A.Account
-instance ToSchema A.InterestInfo
-instance ToSchema F.Fee
-instance ToSchema F.FeeType
-instance ToSchema HE.RateCap
-instance ToSchema LD.Ledger
-instance ToSchema A.ReserveAmount
-instance ToSchema L.Bond
-instance ToSchema L.StepUp
-instance ToSchema L.BondType
-instance ToSchema L.OriginalInfo
-instance ToSchema L.InterestInfo
-instance ToSchema L.InterestOverInterestType
-instance ToSchema (PerPoint (Ratio Integer))
-instance ToSchema (PerCurve Rate)
-instance ToSchema Pre
-instance ToSchema W.PayOrderBy
-instance ToSchema W.ActionWhen
-instance ToSchema W.ExtraSupport
-instance ToSchema W.Action
-instance ToSchema W.BookType
-instance ToSchema W.CollectionRule
-instance ToSchema C.CallOption
-instance ToSchema CE.LiqCreditCalc
-instance ToSchema CE.LiqFacility
-instance ToSchema HE.RateSwap
-instance ToSchema HE.RateSwapType
-instance ToSchema HE.RateSwapBase
-instance ToSchema HE.CurrencySwap
-instance ToSchema CE.LiqSupportType
-instance ToSchema CE.LiqRepayType
-instance ToSchema CE.LiqDrawType
-instance ToSchema CustomDataType
-instance ToSchema TRG.Trigger
-instance ToSchema TRG.TriggerEffect
-instance ToSchema Types.BalanceSheetReport
-instance ToSchema Types.CashflowReport
-instance ToSchema Types.BookItem
-instance ToSchema Stmt.Statement
-instance ToSchema Types.Txn
-instance ToSchema AB.AssociateExp
-instance ToSchema AB.AssociateIncome
-instance ToSchema RV.RevolvingPool
-instance ToSchema (TsPoint [AB.AssetUnion])
-instance ToSchema AP.IssueBondEvent
-instance ToSchema (TsPoint AP.IssueBondEvent)
-instance ToSchema (TsPoint AP.RefiEvent)
-instance ToSchema AP.RefiEvent
-instance ToSchema AP.InspectType
-instance ToSchema AP.CallOpt
-instance ToSchema AP.NonPerfAssumption
-instance ToSchema BondPricingMethod
-instance ToSchema AP.TradeType
-instance ToSchema AP.IrrType
-instance ToSchema AP.BondPricingInput
-instance ToSchema AP.RevolvingAssumption
-instance ToSchema AP.TagMatchRule
-instance ToSchema RangeType
-instance ToSchema AP.FieldMatchRule
-instance ToSchema AP.ObligorStrategy
-instance ToSchema AP.ApplyAssumptionType
-instance ToSchema AP.AssetPerfAssumption
-instance ToSchema AP.AssetDelinqPerfAssumption
-instance ToSchema AP.AssetDefaultedPerfAssumption
-instance ToSchema AP.AssetDefaultAssumption
-instance ToSchema AP.AssetPrepayAssumption
-instance ToSchema AP.RecoveryAssumption
-instance ToSchema RateAssumption
-instance ToSchema AP.ExtraStress
-instance ToSchema AP.AssetDelinquencyAssumption
-instance ToSchema AP.LeaseAssetGapAssump
-instance ToSchema AP.LeaseAssetRentAssump
-instance ToSchema Threshold
-instance ToSchema DB.DealStatFields
-instance ToSchema (PerPoint Balance)
-instance ToSchema (PerCurve Balance)
-instance ToSchema (DB.TestDeal AB.Mortgage)
-instance ToSchema (DB.TestDeal AB.Loan)
-instance ToSchema (DB.TestDeal AB.Installment)
-instance ToSchema (DB.TestDeal AB.Lease)
-instance ToSchema (DB.TestDeal AB.Receivable)
-instance ToSchema (DB.TestDeal AB.ProjectedCashflow)
-instance ToSchema (DB.TestDeal AB.AssetUnion)
-instance ToSchema (DB.TestDeal AB.FixedAsset)
-instance ToSchema (DB.PoolType AB.Mortgage)
-instance ToSchema (DB.PoolType AB.Loan)
-instance ToSchema (DB.PoolType AB.Installment)
-instance ToSchema (DB.PoolType AB.Lease)
-instance ToSchema (DB.PoolType AB.FixedAsset)
-instance ToSchema (DB.PoolType AB.Receivable)
-instance ToSchema (DB.PoolType AB.ProjectedCashflow)
-instance ToSchema (DB.PoolType AB.AssetUnion)
-instance ToSchema (DB.UnderlyingDeal AB.Mortgage)
-instance ToSchema (DB.UnderlyingDeal AB.Loan)
-instance ToSchema (DB.UnderlyingDeal AB.Installment)
-instance ToSchema (DB.UnderlyingDeal AB.Lease)
-instance ToSchema (DB.UnderlyingDeal AB.FixedAsset)
-instance ToSchema (DB.UnderlyingDeal AB.Receivable)
-instance ToSchema (DB.UnderlyingDeal AB.ProjectedCashflow)
-instance ToSchema (DB.UnderlyingDeal AB.AssetUnion)
-instance ToSchema ResultComponent
-instance ToSchema PriceResult
-instance ToSchema DealType
-
-type RunResp = Either String (DealType , Maybe (Map.Map PoolId CF.CashFlowFrame), Maybe [ResultComponent],Map.Map String PriceResult)
 
 wrapRun :: DealType -> Maybe AP.ApplyAssumptionType -> AP.NonPerfAssumption -> RunResp
 wrapRun (MDeal d) mAssump mNonPerfAssump 
@@ -304,21 +139,6 @@ wrapRun (PDeal d) mAssump mNonPerfAssump
 
 wrapRun x _ _ = Left $ "RunDeal Failed ,due to unsupport deal type "++ show x
 
-
-data PoolTypeWrap = LPool (DB.PoolType AB.Loan)
-                  | IPool (DB.PoolType AB.Installment)
-                  | MPool (DB.PoolType AB.Mortgage)
-                  | RPool (DB.PoolType AB.Lease)
-                  | FPool (DB.PoolType AB.FixedAsset)
-                  | VPool (DB.PoolType AB.Receivable)
-                  | PPool (DB.PoolType AB.ProjectedCashflow)
-                  | UPool (DB.PoolType AB.AssetUnion)
-                  deriving(Show, Generic)
-
-
-type RunPoolTypeRtn_ = Map.Map PoolId (CF.CashFlowFrame, Map.Map CutoffFields Balance)
-type RunPoolTypeRtn = Either String RunPoolTypeRtn_
-
 patchCumulativeToPoolRun :: RunPoolTypeRtn_ -> RunPoolTypeRtn_
 patchCumulativeToPoolRun
   = Map.map
@@ -335,13 +155,6 @@ wrapRunPoolType (VPool pt) assump mRates = D.runPoolType pt assump $ Just (AP.No
 wrapRunPoolType (PPool pt) assump mRates = D.runPoolType pt assump $ Just (AP.NonPerfAssumption{AP.interest = mRates})
 wrapRunPoolType (UPool pt) assump mRates = D.runPoolType pt assump $ Just (AP.NonPerfAssumption{AP.interest = mRates})
 wrapRunPoolType x _ _ = Left $ "RunPool Failed ,due to unsupport pool type "++ show x
-
-
-data RunAssetReq = RunAssetReq Date [AB.AssetUnion] (Maybe AP.ApplyAssumptionType) (Maybe [RateAssumption]) (Maybe PricingMethod)
-                   deriving(Show, Generic)
-instance ToSchema RunAssetReq
-
-type RunAssetResp = Either String ((CF.CashFlowFrame, Map.Map CutoffFields Balance), Maybe [PriceResult])
 
 
 wrapRunAsset :: RunAssetReq -> RunAssetResp
@@ -361,66 +174,8 @@ wrapRunAsset (RunAssetReq d assets (Just (AP.PoolLevel assumps)) mRates (Just pm
       let assetCf = P.aggPool Nothing cfs
       return (assetCf , Just pricingResult)
 
-type ScenarioName = String
-
-type DealRunInput = (DealType, Maybe AP.ApplyAssumptionType, AP.NonPerfAssumption)
-
-data RunDealReq = SingleRunReq DealType (Maybe AP.ApplyAssumptionType) AP.NonPerfAssumption
-                | MultiScenarioRunReq DealType (Map.Map ScenarioName AP.ApplyAssumptionType) AP.NonPerfAssumption --- multi pool perf
-                | MultiDealRunReq (Map.Map ScenarioName DealType) (Maybe AP.ApplyAssumptionType) AP.NonPerfAssumption  -- multi deal struct
-                | MultiRunAssumpReq DealType (Maybe AP.ApplyAssumptionType) (Map.Map ScenarioName AP.NonPerfAssumption) -- multi run assump 
-                | MultiComboReq (Map.Map ScenarioName DealType)  (Map.Map ScenarioName (Maybe AP.ApplyAssumptionType))  (Map.Map ScenarioName AP.NonPerfAssumption)
-                deriving(Show, Generic)
-
-data RunSimDealReq = OASReq DealType (Map.Map ScenarioName AP.ApplyAssumptionType) AP.NonPerfAssumption
-                    deriving(Show, Generic)
-
-instance ToSchema PoolTypeWrap
-
-instance ToSchema RunDealReq
-data RunPoolReq = SingleRunPoolReq PoolTypeWrap (Maybe AP.ApplyAssumptionType) (Maybe [RateAssumption])
-                | MultiScenarioRunPoolReq PoolTypeWrap (Map.Map ScenarioName AP.ApplyAssumptionType) (Maybe [RateAssumption])
-                deriving(Show, Generic)
-
-instance ToSchema RunPoolReq
-
-data RunDateReq = RunDateReq Date DatePattern (Maybe Date)
-                deriving(Show, Generic)
-instance ToSchema RunDateReq
-
-
-$(deriveJSON defaultOptions ''DealType)
-$(concat <$> traverse (deriveJSON defaultOptions) [''RunDealReq, ''RunPoolReq,''RunAssetReq, ''RunDateReq,''PoolTypeWrap])
-
 -- Swagger API
 type SwaggerAPI = "swagger.json" :> Get '[JSON] OpenApi
-
-type PoolRunResp = Either String (Map.Map PoolId (CF.CashFlowFrame, Map.Map CutoffFields Balance))
-
-data RootFindResp = FirstLossResult Double AP.ApplyAssumptionType (Maybe AP.RevolvingAssumption)
-                  | BestSpreadResult Double (Map.Map BondName L.Bond) DealType
-                  deriving(Show, Generic)
-
-$(deriveJSON defaultOptions ''RootFindResp)
-instance ToSchema RootFindResp
-
--- type FirstLossResp = Either String FirstLossResult
-
-
-
-type TargetBonds = [BondName]
-
--- calcualte best spread that
---- 1. make sure all bonds are paid off
---- 2. make sure WAC cap is met
-
-data RootFindReq = FirstLossReq DealRunInput BondName
-                 | MaxSpreadToFaceReq DealRunInput (BondName,TargetBonds)
-                 deriving(Show, Generic)
-
-$(deriveJSON defaultOptions ''RootFindReq)
-instance ToSchema RootFindReq
-
 type EngineAPI = "version" :> Get '[JSON] Version
             :<|> "runAsset" :> ReqBody '[JSON] RunAssetReq :> Post '[JSON] RunAssetResp
             :<|> "runPool" :> ReqBody '[JSON] RunPoolReq :> Post '[JSON] PoolRunResp
@@ -433,7 +188,6 @@ type EngineAPI = "version" :> Get '[JSON] Version
             :<|> "runByRootFinder" :> ReqBody '[JSON] RootFindReq :> Post '[JSON] (Either String RootFindResp)
             :<|> "runDate" :> ReqBody '[JSON] RunDateReq :> Post '[JSON] [Date]
 
-
 engineAPI :: Proxy EngineAPI
 engineAPI = Proxy
 
@@ -445,7 +199,6 @@ engineSwagger = toOpenApi engineAPI
                     & info.version .~  T.pack (_version version1)
                     & info.description ?~ "Hastructure is a white-label friendly Cashflow & Analytics Engine for MBS/ABS and REITs"
                     & info.license ?~ "BSD 3"
-
 
 -- showVersion :: Handler (Envelope '[] Version)
 showVersion :: Handler Version
@@ -517,6 +270,17 @@ queryDealType (UDeal _d) = Q.queryCompound _d
 queryDealType (VDeal _d) = Q.queryCompound _d
 queryDealType (PDeal _d) = Q.queryCompound _d
 
+queryDealTypeBool :: DealType -> Date -> DealStats -> Either String Bool
+queryDealTypeBool (MDeal _d) d s = Q.queryDealBool _d s d
+queryDealTypeBool (RDeal _d) d s = Q.queryDealBool _d s d
+queryDealTypeBool (IDeal _d) d s = Q.queryDealBool _d s d
+queryDealTypeBool (LDeal _d) d s = Q.queryDealBool _d s d
+queryDealTypeBool (FDeal _d) d s = Q.queryDealBool _d s d
+queryDealTypeBool (UDeal _d) d s = Q.queryDealBool _d s d
+queryDealTypeBool (VDeal _d) d s = Q.queryDealBool _d s d
+queryDealTypeBool (PDeal _d) d s = Q.queryDealBool _d s d
+
+
 -- stress the pool performance, till a bond suffers first loss
 testByDefault :: DealType -> AP.ApplyAssumptionType -> AP.NonPerfAssumption -> BondName -> Double -> Double
 testByDefault dt assumps nonPerfAssump@AP.NonPerfAssumption{AP.revolving = mRevolving} bn r 
@@ -547,7 +311,8 @@ testBySpread (dt,mPAssump,runAssump) (bn,bnds) f
             v = getPriceValue $ pResult Map.! bn
             bond = dtToBonds d Map.! bn
           in
-            if L.getCurBalance bond > 0 then
+            -- if L.getCurBalance bond > 0 then
+            if True then
               1.0
             else
               (fromRational . toRational) (v - L.getOriginBalance bond)
@@ -638,7 +403,6 @@ instance FromJSON Config
 app :: Application
 app = simpleCors $ serve (Proxy :: Proxy API) myServer
 
-
 main :: IO ()
 main = 
   do 
@@ -650,4 +414,3 @@ main =
                         Right c -> c
     print ("Engine start with version:"++ _version version1++";running at Port:"++ show _p)
     run _p app
--- $(deriveJSON defaultOptions ''DealType)
